+++ guillotina/files.py
@@ -3,24 +3,17 @@ from guillotina import app_settings
 from guillotina import configure
 from guillotina.interfaces import ICloudFileField
 from guillotina.interfaces import IFile
-from guillotina.interfaces import IFileField
 from guillotina.interfaces import IFileManager
 from guillotina.interfaces import IRequest
 from guillotina.interfaces import IResource
 from guillotina.interfaces import IStorage
 from guillotina.interfaces import NotStorable
 from guillotina.utils import import_class
-from persistent import Persistent
-from ZODB.blob import Blob
-from zope.component import adapter
 from zope.component import getMultiAdapter
-from zope.component import getUtility
 from zope.interface import alsoProvides
 from zope.interface import implementer
 from zope.schema import Object
-from zope.schema.fieldproperty import FieldProperty
 
-import aiohttp
 import io
 import mimetypes
 import os
@@ -43,122 +36,6 @@ def get_contenttype(
         return mimetypes.types_map.get(extension, 'application/octet-stream')
 
     return default
-
-
-@configure.adapter(
-    for_=(IResource, IRequest, IFileField),
-    provides=IFileManager)
-class BasicFileManager(object):
-
-    def __init__(self, context, request, field):
-        self.context = context
-        self.request = request
-        self.field = field
-
-    async def upload(self):
-        chunk_size = 8400
-        file = self.field.get(self.field.context)
-        if file is None:
-            filename = None
-            if 'X-UPLOAD-FILENAME' in self.request.headers:
-                filename = self.request.headers['X-UPLOAD-FILENAME']
-            file = BasicFile(filename=filename)
-            self.field.set(self.field.context, file)
-        with file.open('w') as fd:
-            while True:
-                chunk = await self.request.content.read(chunk_size)
-                if not chunk:
-                    break
-                fd.write(chunk)
-
-    async def download(self):
-        file = self.field.get(self.field.context)
-        if file is None:
-            raise AttributeError('No field value')
-
-        resp = aiohttp.web.StreamResponse(headers=aiohttp.MultiDict({
-            'CONTENT-DISPOSITION': 'attachment; filename="%s"' % file.filename
-        }))
-        resp.content_type = file.content_type
-        resp.content_length = file.size
-        await resp.prepare(self.request)
-        resp.write(file.data)
-        await resp.drain()
-        return resp
-
-
-@implementer(IFile)
-class BasicFile(Persistent):
-
-    filename = FieldProperty(IFile['filename'])
-
-    def __init__(self, data='', content_type='', filename=None):
-        if (
-            filename is not None and
-            content_type in ('', 'application/octet-stream')
-        ):
-            content_type = get_contenttype(filename=filename)
-        self.content_type = content_type
-        self._blob = Blob()
-        f = self._blob.open('w')
-        f.write(b'')
-        f.close()
-        self._set_data(data)
-        self.filename = filename
-
-    def open(self, mode='r'):
-        if mode != 'r' and 'size' in self.__dict__:
-            del self.__dict__['size']
-        return self._blob.open(mode)
-
-    def open_detached(self):
-        return open(self._blob.committed(), 'rb')
-
-    def _set_data(self, data):
-        if 'size' in self.__dict__:
-            del self.__dict__['size']
-        # Search for a storable that is able to store the data
-        dotted_name = '.'.join((data.__class__.__module__,
-                                data.__class__.__name__))
-        storable = getUtility(IStorage, name=dotted_name)
-        storable.store(data, self._blob)
-
-    def _get_data(self):
-        fp = self._blob.open('r')
-        data = fp.read()
-        fp.close()
-        return data
-
-    _data = property(_get_data, _set_data)
-    data = property(_get_data, _set_data)
-
-    @property
-    def size(self):
-        if 'size' in self.__dict__:
-            return self.__dict__['size']
-        reader = self._blob.open()
-        reader.seek(0, 2)
-        size = int(reader.tell())
-        reader.close()
-        self.__dict__['size'] = size
-        return size
-
-    def get_size(self):
-        return self.size
-
-
-@implementer(IFileField)
-class BasicFileField(Object):
-    """A NamedBlobFile field
-    """
-
-    _type = BasicFile
-    schema = IFile
-
-    def __init__(self, **kw):
-        if 'schema' in kw:
-            self.schema = kw.pop('schema')
-        super(BasicFileField, self).__init__(schema=self.schema, **kw)
 
 
 @implementer(ICloudFileField)