+++ guillotina/utils.py
@@ -1,16 +1,24 @@
 # -*- coding: utf-8 -*-
 from aiohttp.web_exceptions import HTTPUnauthorized
 from collections import MutableMapping
+from guillotina.exceptions import RequestNotFound
 from hashlib import sha256 as sha
 
 import asyncio
 import fnmatch
 import importlib
+import inspect
 import logging
 import random
 import string
 import sys
 import time
+
+
+try:
+    from aiohttp.web_server import RequestHandler
+except ImportError:
+    from aiohttp.web import RequestHandler
 
 
 try:
@@ -246,3 +254,29 @@ async def apply_coroutine(func, *args, **kwargs):
     if asyncio.iscoroutine(result):
         return await result
     return result
+
+
+def get_current_request():
+    """Return the current request by heuristically looking it up from stack
+    """
+    frame = inspect.currentframe()
+    while frame is not None:
+        request = getattr(frame.f_locals.get('self'), 'request', None)
+        if request is not None:
+            return request
+        elif isinstance(frame.f_locals.get('self'), RequestHandler):
+            return frame.f_locals['request']
+        # if isinstance(frame.f_locals.get('self'), RequestHandler):
+        #     return frame.f_locals.get('self').request
+        # elif IView.providedBy(frame.f_locals.get('self')):
+        #     return frame.f_locals['request']
+        frame = frame.f_back
+    raise RequestNotFound(RequestNotFound.__doc__)
+
+
+try:
+    import guillotina.optimizations  # noqa
+except (ImportError, AttributeError):  # pragma NO COVER PyPy / PURE_PYTHON
+    pass
+else:
+    from guillotina.optimizations import get_current_request  # noqa